+++ sdk/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @outputai/core
 
+## 0.1.8
+
 ## 0.1.7
 
 ### Patch Changes